chore: split out helpers from UnifiedHit
