+++ src/components/command-bar/commands/search/unified-search/components/unified-hit/helpers/index.ts
@@ -1,1 +1,3 @@
 export * from './build-url-path'
+export * from './get-unified-hit-props'
+export * from './parse-default-product-slug'
