+++ src/components/command-bar/commands/search/unified-search/components/unified-hit/types.ts
@@ -1,0 +1,15 @@
+// Components
+// Types
+import type { ProductSlug } from 'types/products'
+
+type AlgoliaContentType = 'docs' | 'tutorial' | 'integration'
+
+export interface UnifiedHitProps {
+	type: AlgoliaContentType
+	href: string
+	ariaLabel: string
+	titleHtml: string
+	descriptionHtml: string
+	productSlug: ProductSlug
+	productName: string
+}
